+++ CHANGES.rst
@@ -27,6 +27,9 @@ Unreleased
   hashes added to their names to ensure updated files are used instead of stale
   cached copies.
 
+- Missing branch coverage explanations that said "the condition was never
+  false" now read "was always true" because it's easier to understand.
+
 
 .. scriv-start-here